preventDefault prevented Events of children
ud-stepper always calls preventDefault on `iron-activate`-Events. This causes a problem when a iron-selector is used inside of a ud-step. We had the problem of a paper-dropdown not being able to get any selection after index 1.
The above code adds a check to see if the iron-activate event actually originates from the ud-stepper itself.