Make stepper accessible
As the stepper inherits from IronMenubarBehavior, it can be navigated
through keyboard.
Role of header is `tablist`, each stepper button has a `tab` role.

When a step is completed, `aria-checked` of its corresponding button
is set to true.

Steps not accessible have `aria-disabled` set to true.

`aria-invalid` is set to true when a step has an error.